rebase

+++ src/main/java/com/mtd/ecom_server/models/Product.java
@@ -3,16 +3,18 @@ package com.mtd.ecom_server.models;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-@Document(collection="products")
+@Document (collection = "products")
 public class Product {
 	@Id
-    private String id;
+	private String id;
 	private String name;
 	private String description;
 	private String category;
 	private String tags;
 	private float price;
-	private float stock;
+	private int stock;
+	
+
 	public String getId() {
 		return id;
 	}
@@ -49,11 +51,11 @@ public class Product {
 	public void setPrice(float price) {
 		this.price = price;
 	}
-	public float getStock() {
+	public int getStock() {
 		return stock;
 	}
-	public void setStock(float stock) {
+	public void setStock(int stock) {
 		this.stock = stock;
 	}
+	
 }
-
